#36632 Move logic back where it belongs
