+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/core/DBeaverActivator.java
@@ -18,40 +18,26 @@ package org.jkiss.dbeaver.core;
 
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.jface.resource.ImageDescriptor;
-import org.eclipse.osgi.service.datalocation.Location;
 import org.eclipse.swt.internal.image.FileFormat;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.jkiss.awt.injector.ProxyInjector;
-import org.jkiss.code.NotNull;
 import org.jkiss.dbeaver.DBeaverPreferences;
-import org.jkiss.dbeaver.Log;
 import org.jkiss.dbeaver.ModelPreferences;
-import org.jkiss.dbeaver.model.app.DBPWorkspace;
 import org.jkiss.dbeaver.model.impl.preferences.BundlePreferenceStore;
 import org.jkiss.dbeaver.model.preferences.DBPPreferenceStore;
 import org.jkiss.dbeaver.model.runtime.features.DBRFeatureRegistry;
 import org.jkiss.dbeaver.ui.ConnectionFeatures;
 import org.jkiss.dbeaver.ui.browser.BrowsePeerMethods;
-import org.jkiss.dbeaver.utils.RuntimeUtils;
 import org.jkiss.utils.ArrayUtils;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
 
 import java.awt.*;
-import java.io.IOException;
 import java.io.PrintStream;
-import java.io.Reader;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.nio.file.FileVisitResult;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.MissingResourceException;
-import java.util.Properties;
 import java.util.ResourceBundle;
-import java.util.Set;
 
 /**
  * The activator class controls the plug-in life cycle
@@ -64,15 +50,6 @@ public class DBeaverActivator extends AbstractUIPlugin {
     private PrintStream debugWriter;
     private DBPPreferenceStore preferences;
 
-    private static final String PLUGINS_FOLDER = ".plugins";
-    private static final String CORE_RESOURCES_PLUGIN_FOLDER = "org.eclipse.core.resources";
-    private static final String STARTUP_ACTIONS_FILE = "dbeaver-startup-actions.properties";
-
-    private static final String RESET_USER_PREFERENCES = "reset_user_preferences";
-    private static final String RESET_WORKSPACE_CONFIGURATION = "reset_workspace_configuration";
-
-    private static final Log log = Log.getLog(DBeaverActivator.class);
-    
     public DBeaverActivator() {
     }
 
@@ -88,7 +65,6 @@ public class DBeaverActivator extends AbstractUIPlugin {
         instance = this;
 
         Bundle bundle = getBundle();
-        resetSettingsStartupActions(Platform.getInstanceLocation());
         ModelPreferences.setMainBundle(bundle);
         preferences = new BundlePreferenceStore(bundle);
 
@@ -206,101 +182,5 @@ public class DBeaverActivator extends AbstractUIPlugin {
 
     public static ImageDescriptor getImageDescriptor(String path) {
         return imageDescriptorFromPlugin(DesktopPlatform.PLUGIN_ID, path);
-    }
-
-    private void resetSettingsStartupActions(@NotNull Location instanceLoc) {
-        Path instancePath;
-        Path actionsPath;
-
-        try {
-            instancePath = RuntimeUtils.getLocalPathFromURL(instanceLoc.getURL()).resolve(DBPWorkspace.METADATA_FOLDER);
-            actionsPath = instancePath.resolve(STARTUP_ACTIONS_FILE);
-        } catch (Exception e) {
-            return;
-        }
-
-        if (Files.notExists(actionsPath)) {
-            return;
-        }
-
-        try (Reader reader = Files.newBufferedReader(actionsPath)) {
-            final Properties properties = new Properties();
-            properties.load(reader);
-
-            if (!properties.isEmpty()) {
-                processResetSettings(instancePath, properties.stringPropertyNames());
-            }
-        } catch (Exception e) {
-            log.error("Unable to read startup actions", e);
-        } finally {
-            try {
-                Files.delete(actionsPath);
-            } catch (IOException e) {
-                log.error("Unable to delete startup actions file: " + e.getMessage());
-            }
-        }
-    }
-
-    private void processResetSettings(
-        @NotNull Path instancePath,
-        @NotNull Set<String> actions
-    ) throws Exception {
-        final boolean resetUserPreferences = actions.contains(RESET_USER_PREFERENCES);
-        final boolean resetWorkspaceConfiguration = actions.contains(RESET_WORKSPACE_CONFIGURATION);
-
-        if (!resetUserPreferences && !resetWorkspaceConfiguration) {
-            return;
-        }
-        Path path = instancePath.resolve(PLUGINS_FOLDER);
-        if (Files.notExists(path) || !Files.isDirectory(path)) {
-            return;
-        }
-
-        Files.walkFileTree(path, new SimpleFileVisitor<>() {
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) {
-                log.trace("Deleting " + file);
-
-                try {
-                    Files.delete(file);
-                } catch (IOException e) {
-                    log.trace("Unable to delete " + file + ":" + e.getMessage());
-                }
-
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) {
-                if (dir.endsWith(PLUGINS_FOLDER)) {
-                    return FileVisitResult.CONTINUE;
-                }
-
-                final Path relative = path.relativize(dir);
-
-                if (resetUserPreferences && !relative.startsWith(CORE_RESOURCES_PLUGIN_FOLDER)) {
-                    return FileVisitResult.CONTINUE;
-                }
-
-                if (resetWorkspaceConfiguration && relative.startsWith(CORE_RESOURCES_PLUGIN_FOLDER)) {
-                    return FileVisitResult.CONTINUE;
-                }
-
-                return FileVisitResult.SKIP_SUBTREE;
-            }
-
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException exc) {
-                log.trace("Deleting " + dir);
-
-                try {
-                    Files.delete(dir);
-                } catch (IOException e) {
-                    log.trace("Unable to delete " + dir + ":" + e.getMessage());
-                }
-
-                return FileVisitResult.CONTINUE;
-            }
-        });
     }
 }
